requirements.txt
pillow
django-admin-interface
crispy-bootstrap5
